// src/app/Interfaces/assetStatus.ts
export class AssetStatus {
    // assetName: string;
    // assetType: string;
    // assetState: string;
    // assetLocation: string;
    // timestamp: string;
    // bleStatus: string;
    // trackerStatus: string;

    AssetID: number;
    AssetName: string;
    AssetType: string;
    AssetSerialNo: string;
    AssetTagSerialNo: string;
    Status: string;
    AssetLocation: string;
    StartDateTime: string;
    BleStatus: string;
    TrackerStatus: string;
    User: string;
    AssetImageUrl: string;
    LastUpdatedtDateTime: string;
    AssetTrackerNo: string;
    AssetCategory: string;
    IndoorAlertsCount: number;
    Temperature:number;
    Humidity:number;
}